feat(mergedmining): accept miner login indicator for nicehash mindiff
NiceHash [requires pools to set a minimum mining difficulty of
500000](https://www.nicehash.com/pool-operators). This doesn't work well
for individual workers (i.e. an Antminer S9) without a proxy (for
various reasons, but mostly because we need shares to be submitted
quickly to keep orphan blocks at a minimum) so the public merged-mining
server uses a much lower minimum (and initial) difficulty. Which leaves
off NiceHash miners. However with this patch it is possible for miners
to include the string `nicehash` (in any capitalization) somewhere in
their password to trigger the NiceHash compatibility. Which should
include any other quirk that comes up.