chore(docker): fix build, broke due to recent updates
Recent updates on the cryptography package introduced the need for rustc
for compiling it from source. Although the required runtime version of
cryptography is locked in `poetry.lock`, when installing Poetry (from
outside the controlled virtualenv) it gets pulled by some dependency and
the latest version is installed. The bare `Dockerfile` has been updated
to support these changes (install Rust and update alpine).

The last update to the workflow script that builds and pushes the image
stopped adding the `latest` tag. I ported the script to Python because
it's better to maintain than the Bash script we had. This update also
includes some small adjustments to support easier testing of this build
script and use the Python version on the `Dockerfile`s as the default
(as opposed to encoding it in the script).