refactor(settings): make HathorSettings always return the same instance
This wasn't _exactly_ the case before this commit. On one side when
using the deprecated module settings the same module would always be
loaded, and the same `SETTINGS` property would be accessed, but
reloading the module (or at least leaving that up to importlib)
unnecessary. On the new YAML system the file is reloaded everytime, this
is also unnecessary.

This commit will make it the last HathorSettings instantiated will be
returned, as long as the source is the same, otherwise an exception is
raised.

There is a little small caveat that is fixed in this commit is that in
the YAML file could in theory be altered between different calls to
HathorSettings, which would lead to different content being loaded
silently. What happens now is that the first time it is loaded is what
is used for every call, any change in the file won't have any effect.