feat(cli): add argument to make sync-v1 unavailable
Currently it is possibly to disable sync-v1 but not make it unavailable.
Making it unavailble will be required in order to be able to disable the
indexes that it needs.

This commit does not bother on any of the effects that are possible when
sync-v1 is unavailable, only on making it possible to do so via a
command line argument.