chore: update poetry lock file so python-rocksdb is v0.9.3
Even if #1244 has updated the resolved_reference, it seems it didn't do
a complete update (python-rocksdb version was still marked as 0.9.2
instead of 0.9.3). The lock file was generated with poetry-2.1.1, and
thus includes some new fields.

The only downside of not doing this change seems to be that in
development setups where python-rocksdb v0.9.2 exists in the cache, it
will be used instead of correctly installing the one from github with
version v0.9.3.